feat: search trips
Helps user to plan a journey

- page: search origin
- page: search destination
- page: list trips (with details)

This is for people like me who likes to check commute distance very frequently.

While writing this, a lot of code was duplicated from the original `SearchPage`. There is a serious need for refactor. I didn't want to do it straightaway and risk touching other parts of the code as well. However, if this feature is merged, we could ideally create a standalone generic enough search page which:
- only searches for venue
- gets Primary action from parent,
- gets all other actions from a common file (so all commands can have the same common actions for consistency)
- possibly use inbuilt throttle of `List`

resolves #10