feat: prioritize trip transport mode accessories instead of duration
The list of transport mode icons and public code should be a decent
visualization of the kind and complexity of a trip suggestion. When a
lot of trips are listed after another, I also assume it's a more
effective way to compare options at a glance than the duration.

It does however make it hard to fit everything for complex trips, so
there's some logic here to hide the publicCode and truncate the list
when necessary.

To also make the duration available, it's added to the last item in the
details list.