chore: remove pagination TODO
It seems that `@raycast/utils` useFetch won't support pagination with
GraphQL, given the way it implements cursors. This is because the cursor
is provided as an attribute when constructing the URL instead of when
building the body.

That is, it supports

`(options) => 'https://example.com?cursor=' + options.cursor`

but not

`body: JSON.stringify({variables: { cursor: options.cursor } })`

https://developers.raycast.com/utilities/react-hooks/usefetch